Control.Monad.Exception.Throws

Description

Defines the Throws binary relationship between types.

Documentation

class Exception e => Throws e l

Throws is a type level binary relationship used to model a list of exceptions.

There is only one case in which the user must add further instances to Throws. If your sets of exceptions are hierarchical then you need to teach Throws about the hierarchy.

Subtyping

As there is no way to automatically infer the subcases of an exception, they have to be encoded manually mirroring the hierarchy defined in the defined Exception instances. For example, the following instance encodes that MyFileNotFoundException is a subexception of MyIOException :

 instance Throws MyFileNotFoundException (Caught MyIOException l)

Throws is not a transitive relation and every ancestor relation must be explicitly encoded.

                                                            --   TopException
                                                            --         |
   instance Throws MidException   (Caught TopException l)   --         |
                                                            --   MidException
   instance Throws ChildException (Caught MidException l)   --         |
   instance Throws ChildException (Caught TopException l)   --         |
                                                            --  ChildException

Note that SomeException is automatically an ancestor of every other exception type.

data Caught e l

A type level witness of a exception handler.

newtype CheckedException l

CheckedException adds a phantom type parameter l to SomeException

Constructors

CheckedException

Fields

checkedException :: SomeException